requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
weaviate-client==3.24.1
